Bump RTT version
This uses the SEGGER RTT files included with SEGGER SystemView v2.32.
Whitespace changes are observed only, so technically we'd be able to
just use the old files, but I'm doing this bump regardless for traceability.

IMPORTANT: I guess SEGGER SystemView and RTT should be upgraded in
lockstep with each other and not separately, since RTT is supplied with
the SystemView source package. The separation in different
dirs is kept in order to prevent any breakage.

In any case, if updates are ever needed, *prefer strongly to use the
RTT files that come with SystemView and upgrade both at the same time.

Attempting to use a different RTT version may cause conflicts that
are a nightmare to debug.