Add printf() implementation by SEGGER RTT
From now on, if printf() is called in your C programs and this file is
built, calls to it will be redirected to RTT instead.

This might not be actually needed; still adding it and we can remove
it if the necessity arises - it's just a nice to have that's given
to us.